requirements/requirements_yolo.txt
huggingface_hub==1.3.5
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.13.0.90
pillow==12.1.0
PyYAML==6.0.3
scipy==1.15.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==5.0.0
ultralytics==8.4.8

